requirements.txt
Django==5.0
django-browser-reload==1.12.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-htmx==1.17.2
django-mptt==0.15.0
django-template-partials==23.4
django-types==0.19.1
extruct==0.16.0
factory-boy==3.3.0
gunicorn==21.2.0
lxml[html_clean]==4.9.3.
Markdown==3.5.1
nh3==0.2.15
Pygments==2.17.2
pymdown-extensions==10.5
django-compressor
django-redis
django-clickhouse-backend
celery
